apps/murmur-rest/requirements.txt
flask==3.1.0
zeroc-ice==3.7.10
gunicorn==23.0.0
